bash: ignore shellcheck warning SC2124
Shellcheck emits a warning:-
SC2124: Assigning an array to a string! Assign as array, or use * instead of @ to concatenate.

However, in bash the contents of $@ is concatenated with s separating
space, this is then fine as an IFS separator on the for x in $y loop.
So, silence the warning.

Link: https://www.shellcheck.net/wiki/SC2124
Signed-off-by: Mark Grant <m.grant.prg@gmail.com>